Move null output stream in test utilities Add unit tests
