+++ node/silkworm/test/log.hpp
@@ -16,6 +16,8 @@
 
 #pragma once
 
+#include <ostream>
+
 #include <silkworm/common/log.hpp>
 
 namespace silkworm::test {
@@ -31,5 +33,16 @@ class SetLogVerbosityGuard {
   private:
     log::Level current_level_;
 };
+
+// Factory function creating one null output stream (all characters are discarded)
+inline std::ostream& null_stream() {
+    static struct null_buf : public std::streambuf {
+        int overflow(int c) override { return c; }
+    } null_buf;
+    static struct null_strm : public std::ostream {
+        null_strm() : std::ostream(&null_buf) {}
+    } null_strm;
+    return null_strm;
+}
 
 }  // namespace silkworm::test
